update route rules
