+++ src/main.tsx
@@ -2,7 +2,7 @@
 import React, { Suspense } from 'react';
 import ReactDOM from 'react-dom';
 import { Route } from 'react-router';
-import { BrowserRouter, Routes } from 'react-router-dom';
+import { BrowserRouter, Routes, Navigate } from 'react-router-dom';
 
 //redux 相关
 import { Provider } from 'react-redux';
@@ -16,6 +16,8 @@ import '@/i18n';
 
 //视图与组件
 import NotFound from '@/views/NotFound';
+import Home from './views/Home';
+import { Resource } from '@/views/Resource';
 import DataCenter from '@/views/DataCenter';
 import Dashboard from './views/Dashboard';
 import LoginPage from '@/views/Login';
@@ -24,19 +26,20 @@ import Event from '@/views/Event';
 import { CFullLoading } from '@/components/Common/CFullLoading';
 import { CHeader } from './components/Logic/CHeader';
 import { CFooter } from './components/Logic/CFooter';
-import { Home } from '@/views/Home';
 
 
 const AppRouter = (): JSX.Element => {
     return(<>
         <CHeader/>
         <Routes>
-            <Route path="dataCenter" element={<DataCenter />} />
+            <Route path="dataCenter/*" element={<DataCenter />} />
             <Route path="home/*" element={<Home />} />
-            <Route path="dashboard" element={<Dashboard />} />
-            {/* <Route path="404" element={<NotFound />} /> */}
+            <Route path="dashboard/*" element={<Dashboard />} />
             <Route path="account/*" element={<Account />} />
-            <Route path="event" element={<Event />} />
+            <Route path="account" element={<Navigate to="/account/profile" />} />
+            <Route path="event/*" element={<Event />} />
+            <Route path="resource/*" element={<Resource />} />
+            <Route path="resource" element={<Navigate to="/resource/server" />} />
             <Route path="*" element={<NotFound />} />
         </Routes>
         <CFooter />
