+++ src/views/Resource/Server/AddServer/DiskConfiguration.tsx
