+++ src/views/Resource/Server/ServerDetail/Detail.tsx
