+++ src/views/Resource/Server/ServerDetail/Network.tsx
