+++ src/views/Resource/Server/ServerDetail/index.tsx
